third_party/WebKit/LayoutTests/imported/wpt/html/browsers/browsing-the-web/navigating-across-documents/002-expected.txt

This is a testharness.js-based test.
FAIL Multiple simultaneous navigations assert_equals: expected "002-2" but got "002-1"
Harness: the test ran to completion.
